@@ -1,12 +1,9 @@
-﻿using System;
+﻿namespace street_service.Exceptions;
 
-namespace street_service.Exceptions
+public class ProviderException : Exception
 {
-    public class ProviderException : Exception
+    public ProviderException(string message)
+        : base(message)
     {
-        public ProviderException(string message)
-            : base(message)
-        {
-        }
     }
 }
