@@ -1,26 +1,16 @@
-﻿using StockportGovUK.NetStandard.Gateways.VerintServiceGateway;
-using StockportGovUK.NetStandard.Models.Addresses;
-using StockportGovUK.NetStandard.Models.Enums;
-using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿namespace street_service.Providers;
 
-namespace street_service.Providers
+public class VerintStreetProvider : IStreetProvider
 {
-    public class VerintStreetProvider : IStreetProvider
+    public EStreetProvider ProviderName => EStreetProvider.CRM;
+
+    private readonly IVerintServiceGateway _verintServiceGateway;
+
+    public VerintStreetProvider(IVerintServiceGateway verintServiceGateway) => _verintServiceGateway = verintServiceGateway;
+
+    public async Task<IEnumerable<AddressSearchResult>> SearchAsync(string street)
     {
-        public EStreetProvider ProviderName => EStreetProvider.CRM;
-
-        private readonly IVerintServiceGateway _verintServiceGateway;
-
-        public VerintStreetProvider(IVerintServiceGateway verintServiceGateway)
-        {
-            _verintServiceGateway = verintServiceGateway;
-        }
-
-        public async Task<IEnumerable<AddressSearchResult>> SearchAsync(string street)
-        {
-            var response = await _verintServiceGateway.GetStreetByReference(street);
-            return response.ResponseContent;
-        }
+        var response = await _verintServiceGateway.GetStreetByReference(street);
+        return response.ResponseContent;
     }
 }
