@@ -1,15 +1,9 @@
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
+namespace street_service.Utils.HealthChecks;
 
-namespace street_service.Utils.HealthChecks
+public class TestHealthCheck : IHealthCheck
 {
-    public class TestHealthCheck : IHealthCheck
+    public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
     {
-        public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
-        {
-            return HealthCheckResult.Healthy(null, new Dictionary<string, object> {{"Result", "All working!"}});
-        }
+        return await Task.FromResult(HealthCheckResult.Healthy(null, new Dictionary<string, object> { { "Result", "All working!" } }));
     }
 }
