@@ -1,9 +1,6 @@
-using System;
+namespace street_service.Utils.Toggles;
 
-namespace street_service.Utils.Toggles
+public static class OperationalToggles
 {
-    public static class OperationalToggles
-    {
-        public const string MyToggle = "MyToggle";
-    }
+    public const string MyToggle = "MyToggle";
 }
